refactor: juntando os tipos de músicas
