--- src/shared/types/MusicaAtual.types.ts
@@ -1,11 +1,0 @@
-export interface MusicaAtual {
-  id: number;
-  imagemURL: string;
-  nomeMusica: string;
-  nomeArtista: string;
-}
-
-export interface MusicaAtualContextType {
-  musicaAtual: MusicaAtual | null;
-  setMusicaAtual: (musica: MusicaAtual | null) => void;
-}
