--- src/shared/types/Playlist.types.ts
@@ -1,9 +1,0 @@
-import { MusicaType } from "./Musica.types";
-
-export type PlaylistType = {
-  titulo: string;
-  descricao: string;
-  autor: string;
-  imagem: string;
-  musicas: MusicaType[];
-};
