+++ src/shared/types/musica.types.ts
@@ -1,0 +1,44 @@
+// ---- Backend ----
+
+export interface Tag {
+  id: string;
+  nome: string;
+}
+
+export interface GeneroMusical {
+  id: string;
+  nome: string;
+}
+
+export interface MusicaBackend {
+  id: string;
+  nome: string;
+  caminhoDoArquivo: string;
+  letra: string;
+  tags: Tag[];
+  generosMusicais: GeneroMusical[];
+}
+
+// ---- Frontend / Provider ----
+
+export interface MusicaAtual {
+  id: string;
+  imagemURL: string;
+  nomeMusica: string;
+  nomeArtista: string;
+}
+
+export interface MusicaType extends MusicaAtual {
+  album: string;
+  duracao: string;
+}
+
+// ---- Playlist ----
+
+export type PlaylistType = {
+  titulo: string;
+  descricao: string;
+  autor: string;
+  imagem: string;
+  musicas: MusicaType[];
+};
